Upstream version 20200619+dfsg.1

+++ nntp/browse.c
@@ -27,6 +27,7 @@
  */
 
 #include "config.h"
+#include <stdint.h>
 #include <stdio.h>
 #include "mutt/lib.h"
 #include "email/lib.h"
@@ -51,7 +52,7 @@
 const char *group_index_format_str(char *buf, size_t buflen, size_t col, int cols,
                                    char op, const char *src, const char *prec,
                                    const char *if_str, const char *else_str,
-                                   unsigned long data, MuttFormatFlags flags)
+                                   intptr_t data, MuttFormatFlags flags)
 {
   char fn[128], fmt[128];
   struct Folder *folder = (struct Folder *) data;
